GoF {ltm}	R Documentation

Goodness of Fit for Rasch Models

Description

Performs a parametric Bootstrap test for Rasch and Generalized Partial Credit models.

Usage

GoF.gpcm(object, simulate.p.value = TRUE, B = 99, seed = NULL, ...)

GoF.rasch(object, B = 49, ...)

Arguments

object
an object inheriting from either class gpcm or class rasch.

simulate.p.value
logical; if TRUE, the reported p-value is based on a parametric Bootstrap approach. Otherwise the p-value is based on the asymptotic chi-squared distribution.

B
the number of Bootstrap samples. See Details section for more info.

seed
the seed to be used during the parametric Bootstrap; if NULL, a random seed is used.

...
additional arguments; currently none is used.

Details

GoF.gpcm and GoF.rasch perform a parametric Bootstrap test based on Pearson's chi-squared statistic defined as

∑_{r=1}^{2^p} (O_r - E_r)^2 / E_r,

where r represents a response pattern, O_r and E_r represent the observed and expected frequencies, respectively and p denotes the number of items. The Bootstrap approximation to the reference distribution is preferable compared with the ordinary Chi-squared approximation since the latter is not valid especially for large number of items (=> many response patterns with expected frequencies smaller than 1).

In particular, the Bootstrap test is implemented as follows:

Step 0:

Based on object compute the observed value of the statistic T_{obs}.

Step 1:

Simulate new parameter values, say θ^*, from N(\hat{θ}, C(\hat{θ})), where \hat{θ} are the MLEs and C(\hat{θ}) their large sample covariance matrix.

Step 2:

Using θ^* simulate new data (with the same dimensions as the observed ones), fit the generalized partial credit or the Rasch model and based on this fit calculate the value of the statistic T_i.

Step 3:

Repeat steps 1-2 B times and estimate the p-value using [1 + {\# T_i > T_{obs}}]/(B + 1).

Furthermore, in GoF.gpcm when simulate.p.value = FALSE, then the p-value is based on the asymptotic chi-squared distribution.

Value

An object of class GoF.gpcm or GoF.rasch with components,

Tobs
the value of the Pearson's chi-squared statistic for the observed data.

B
the B argument specifying the number of Bootstrap samples used.

call
the matched call of object.

p.value
the p-value of the test.

simulate.p.value
the value of simulate.p.value argument (returned on for class GoF.gpcm).

df
the degrees of freedom for the asymptotic chi-squared distribution (returned on for class GoF.gpcm).

Author(s)

Dimitris Rizopoulos d.rizopoulos@erasmusmc.nl

See Also

person.fit, item.fit, margins, gpcm, rasch

Examples

## GoF for the Rasch model for the LSAT data:
fit <- rasch(LSAT)
GoF.rasch(fit)

[Package ltm version 1.1-1 Index]