cronbach.alpha {ltm}	R Documentation

Cronbach's alpha

Description

Computes Cronbach's alpha for a given data-set.

Usage

cronbach.alpha(data, standardized = FALSE, CI = FALSE, 
    probs = c(0.025, 0.975), B = 1000, na.rm = FALSE)

Arguments

data
a matrix or a data.frame containing the items as columns.

standardized
logical; if TRUE the standardized Cronbach's alpha is computed.

CI
logical; if TRUE a Bootstrap confidence interval for Cronbach's alpha is computed.

probs
a numeric vector of length two indicating which quantiles to use for the Bootstrap CI.

B
the number of Bootstrap samples to use.

na.rm
logical; what to do with NA's.

Details

The Cronbach's alpha computed by cronbach.alpha() is defined as follows

alpha = (p / (p - 1)) (1 - (∑_{i=1}^p sigma_{y_i}^2/ sigma_x^2)),

where p is the number of items sigma_x^2 is the variance of the observed total test scores, and sigma_{y_i}^2 is the variance of the ith item.

The standardized Cronbach's alpha computed by cronbach.alpha() is defined as follows

alpha_s = (p r) / (1 + (p - 1) r),

where p is the number of items, and r is the average of all (Pearson) correlation coefficients between the items. In this case if na.rm = TRUE, then the complete observations (i.e., rows) are used.

The Bootstrap confidence interval is calculated by simply taking B samples with replacement from data, calculating for each alpha or alpha_s, and computing the quantiles according to probs.

Value

cronbach.alpha() returns an object of class cronbachAlpha with components

alpha
the value of Cronbach's alpha.

n
the number of sample units.

p
the number of items.

standardized
a copy of the standardized argument.

name
the name of argument data.

ci
the confidence interval for alpha; returned if CI = TRUE.

probs
a copy of the probs argument; returned if CI = TRUE.

B
a copy of the B argument; returned if CI = TRUE.

Author(s)

Dimitris Rizopoulos d.rizopoulos@erasmusmc.nl

References

Cronbach, L. J. (1951) Coefficient alpha and the internal structure of tests. Psychometrika, 16, 297–334.

Examples

# Cronbach's alpha for the LSAT data-set
# with a Bootstrap 95% CI
cronbach.alpha(LSAT, CI = TRUE, B = 500)

[Package ltm version 1.1-1 Index]