information {ltm}	R Documentation

Area under the Test or Item Information Curves

Description

Computes the amount of test or item information for a fitted IRT model, in a specified range.

Usage

information(object, range, items = NULL, ...)

Arguments

object
an object inheriting from either class gpcm, class grm, class ltm, class rasch or class tpm.

range
a numeric interval for which the test information should be computed.

items
the items for which the information should be computed; the default NULL corresponds to all the items, which is equivalent to the test information.

...
extra arguments passed to integrate().

Details

The amount of information is computed as the area under the Item or Test Information Curve in the specified interval, using integrate().

Value

A list of class information with components,

InfoRange
the amount of information in the specified interval.

InfoTotal
the total amount of information; typically this is computed as the amount of information in the interval (-10, 10).

PropRange
the proportion of information in the specified range, i.e., "Info in range" / "Total Info".

range
the value of range argument.

items
the value of items argument.

call
the matched call for object.

Author(s)

Dimitris Rizopoulos d.rizopoulos@erasmusmc.nl

See Also

plot.gpcm, plot.grm, plot.ltm, plot.rasch

Examples

fit <- rasch(LSAT)
information(fit, c(-2, 0))
information(fit, c(0, 2), items = c(3, 5))

[Package ltm version 1.1-1 Index]