item.fit {ltm}	R Documentation

Item-Fit Statistics and P-values

Description

Computation of item fit statistics for ltm, rasch and tpm models.

Usage

item.fit(object, G = 10, FUN = median, 
         simulate.p.value = FALSE, B = 100)

Arguments

object
a model object inheriting either from class ltm, class rasch or class tpm.

G
either a number or a numeric vector. If a number, then it denotes the number of categories sample units are grouped according to their ability estimates.

FUN
a function to summarize the ability estimate with each group (e.g., median, mean, etc.).

simulate.p.value
logical; if TRUE, then the Monte Carlo procedure described in the Details section is used to approximate the the distribution of the item-fit statistic under the null hypothesis.

B
the number of replications in the Monte Carlo procedure.

Details

The item-fit statistic computed by item.fit() has the form:

∑_{j = 1}^g N_j (O_{ij} - E_{ij})^2 / [E_{ij} (1 - E_{ij})],

where i is the item, j is the interval created by grouping sample units on the basis of their ability estimates, G is the number of sample units groupings (i.e., G argument), N_j is the number of sample units with ability estimates falling in a given interval j, O_{ij} is the observed proportion of keyed responses on item i for interval j, and E_{ij} is the expected proportion of keyed responses on item i for interval j based on the IRT model (i.e., object) evaluated at the ability estimate z^* within the interval, with z^* denoting the result of FUN applied to the ability estimates in group j.

If simulate.p.value = FALSE, then the p-values are computed assuming a chi-squared distribution with degrees of freedom equal to the number of groups G minus the number of estimated parameters. If simulate.p.value = TRUE, a Monte Carlo procedure is used to approximate the distribution of the item-fit statistic under the null hypothesis. In particular, the following steps are replicated B times:

Step 1:

Simulate a new data-set of dichotomous responses under the assumed IRT model, using the maximum likelihood estimates \hat{theta} in the original data-set, extracted from object.

Step 2:

Fit the model to the simulated data-set, extract the maximum likelihood estimates theta^* and compute the ability estimates z^* for each response pattern.

Step 3:

For the new data-set, and using z^* and theta^*, compute the value of the item-fit statistic.

Denote by T_{obs} the value of the item-fit statistic for the original data-set. Then the p-value is approximated according to the formula

(1 + sum_{b = 1}^B I(T_b >= T_{obs})) / (1 + B),

where I(.) denotes the indicator function, and T_b denotes the value of the item-fit statistic in the bth simulated data-set.

Value

An object of class itemFit is a list with components,

Tobs
a numeric vector with item-fit statistics.

p.values
a numeric vector with the corresponding p-values.

G
the value of the G argument.

simulate.p.value
the value of the simulate.p.value argument.

B
the value of the B argument.

call
a copy of the matched call of object.

Author(s)

Dimitris Rizopoulos d.rizopoulos@erasmusmc.nl

References

Reise, S. (1990) A comparison of item- and person-fit methods of assessing model-data fit in IRT. Applied Psychological Measurement, 14, 127–137.

Yen, W. (1981) Using simulation results to choose a latent trait model. Applied Psychological Measurement, 5, 245–262.

See Also

person.fit, margins, GoF.gpcm, GoF.rasch

Examples

# item-fit statistics for the Rasch model
# for the Abortion data-set
item.fit(rasch(Abortion))

# Yen's Q1 item-fit statistic (i.e., 10 latent ability groups; the
# mean ability in each group is used to compute fitted proportions) 
# for the two-parameter logistic model for the LSAT data-set
item.fit(ltm(LSAT ~ z1), FUN = mean)

[Package ltm version 1.1-1 Index]