ltm-package {ltm}	R Documentation

Latent Trait Models for Item Response Theory Analyses

Description

This package provides a flexible framework for Item Response Theory analyses for dichotomous and polytomous data under a Marginal Maximum Likelihood approach. The fitting algorithms provide valid inferences under Missing At Random missing data mechanisms.

Details

Package: 	 ltm

Type: 	 Package

Version: 	 1.1-1

Date: 	 2018-04-17

License: 	 GPL

The following options are available:

Descriptives:

samples proportions, missing values information, biserial correlation of items with total score, pairwise associations between items, Cronbach's alpha, unidimensionality check using modified parallel analysis, nonparametric correlation coefficient, plotting of sample proportions versus total score.

Dichotomous data:

Rasch Model, Two Parameter Logistic Model, Birnbaum's Three Parameter Model, and Latent Trait Model up to two latent variables (allowing also for nonlinear terms between the latent traits).

Polytomous data:

Samejima's Graded Response Model and the Generalized Partial Credit Model.

Goodness-of-Fit:

Bootstrapped Pearson χ^2 for Rasch and Generalized Partial Credit models, fit on the two- and three-way margins for all models, likelihood ratio tests between nested models (including AIC and BIC criteria values), and item- and person-fit statistics.

Factor Scoring - Ability Estimates:

Empirical Bayes (i.e., posterior modes), Expected a posteriori (i.e., posterior means), Multiple Imputed Empirical Bayes, and Component Scores for dichotomous data.

Test Equating:

Alternate Form Equating (where common and unique items are analyzed simultaneously) and Across Sample Equating (where different sets of unique items are analyzed separately based on previously calibrated anchor items).

Plotting:

Item Characteristic Curves, Item Information Curves, Test Information Functions, Standard Error of Measurement, Standardized Loadings Scatterplot (for the two-factor latent trait model), Item Operation Characteristic Curves (for ordinal polytomous data), Item Person Maps.

Author(s)

Dimitris Rizopoulos

Maintainer: Dimitris Rizopoulos <d.rizopoulos@erasmusmc.nl>

References

Baker, F. and Kim, S-H. (2004) Item Response Theory, 2nd ed. New York: Marcel Dekker.

Rizopoulos, D. (2006) ltm: An R package for latent variable modelling and item response theory analyses. Journal of Statistical Software, 17(5), 1–25. URL http://www.jstatsoft.org/v17/i05/

[Package ltm version 1.1-1 Index]