plot descript {ltm}	R Documentation

Descriptive Statistics Plot method

Description

The plot method for descript objects currently works for dichotomous response patterns, and produces the xy-plot of the total score versus the proportion of correct responses for each item.

Usage

## S3 method for class 'descript'
plot(x, items = NULL, includeFirstLast = FALSE, xlab, ylab, ...)

Arguments

x
an object inheriting from class descript.

items
a numeric vector indicating which items to plot.

includeFirstLast
logical; if TRUE the first and last total scores categories are included.

xlab, ylab
character string or an expression; see title.

...
extra graphical parameters to be passed to matplot().

Author(s)

Dimitris Rizopoulos d.rizopoulos@erasmusmc.nl

See Also

descript

Examples

## Descriptives for WIRS data:
dsc <- descript(WIRS)
dsc
plot(dsc, includeFirstLast = TRUE, type = "b", lty = 1, pch = 1:6)
legend("topleft", names(WIRS), pch = 1:6, col = 1:6, lty = 1, bty = "n")

[Package ltm version 1.1-1 Index]