plot fscores {ltm}	R Documentation

Factor Scores - Ability Estimates Plot method

Description

Plots a Kernel Density Estimation of the distribution of the factor scores (i.e., person parameters). Provides also the option to include in the plot the item difficulty parameters (similar to the Item Person Maps).

Usage

## S3 method for class 'fscores'
plot(x, bw = "nrd0", adjust = 2, kernel = "gaussian", 
    include.items = FALSE, tol = 0.2, xlab = "Ability", ylab = "Density", 
    main = "Kernel Density Estimation for Ability Estimates", 
    pch = 16, cex = 1.5, ...)

Arguments

x
an object inheriting from class fscores.

bw, adjust, kernel
arguments to density().

include.items
logical; if TRUE the item difficulty parameters are included in the plot.

tol
the tolerance used to group the item difficulty parameters, i.e., when include.items = TRUE the values round(betas / tol) * tol are plotted, where beta is the numeric vector of item difficulty parameters.

xlab, ylab, main
character string or an expression; see title.

pch, cex
arguments to stripchart(); used when include.items = TRUE.

...
extra graphical parameters to be passed to plot.density().

Author(s)

Dimitris Rizopoulos d.rizopoulos@erasmusmc.nl

See Also

factor.scores

Examples

## Factor Scores for LSAT data:
fsc <- factor.scores(rasch(LSAT))
plot(fsc, include.items = TRUE, main = "KDE for Person Parameters")
legend("left", "item parameters", pch = 16, cex = 1.5, bty = "n")

[Package ltm version 1.1-1 Index]