rasch {ltm}	R Documentation

Rasch Model

Description

Fit the Rasch model under the Item Response Theory approach.

Usage

rasch(data, constraint = NULL, IRT.param = TRUE, start.val = NULL, 
    na.action = NULL, control = list(), Hessian = TRUE)

Arguments

data
a data.frame (that will be converted to a numeric matrix using data.matrix()) or a numeric matrix of manifest variables.

constraint
a two-column numeric matrix with at most p rows (where p is the number of items), specifying fixed-value constraints. The first column represents the item (i.e., 1 denotes the first item, 2 the second, etc., and p+1 the discrimination parameter) and the second column the value at which the corresponding parameter should be fixed. See Examples for more info.

IRT.param
logical; if TRUE then the coefficients' estimates are reported under the usual IRT parameterization. See Details for more info.

start.val
the character string "random" or a numeric vector of p+1 starting values, where the first p values correspond to the easiness parameters while the last value corresponds to the discrimination parameter. If "random", random starting values are used. If NULL starting values are automatically computed.

na.action
the na.action to be used on data. In case of missing data, if na.action = NULL the model uses the available cases, i.e., it takes into account the observed part of sample units with missing values (valid under MAR mechanisms if the model is correctly specified). If you want to apply a complete case analysis then use na.action = na.exclude.

control
a list of control values,

iter.qN

the number of quasi-Newton iterations. Default 150.

GHk

the number of Gauss-Hermite quadrature points. Default 21.

method

the optimization method to be used in optim(). Default "BFGS".

verbose

logical; if TRUE info about the optimization procedure are printed.

Hessian
logical; if TRUE, then the Hessian matrix is computed. Warning: setting this argument to FALSE will cause many methods (e.g., summary()) to fail; setting to FALSE is intended for simulation purposes in order rasch() to run faster.

Details

The Rasch model is a special case of the unidimensional latent trait model when all the discrimination parameters are equal. This model was first discussed by Rasch (1960) and it is mainly used in educational testing where the aim is to study the abilities of a particular set of individuals.

The model is defined as follows

logit (π_i) = beta_{i} + beta z,

where π_i denotes the conditional probability of responding correctly to the ith item given z, beta_{i} is the easiness parameter for the ith item, β is the discrimination parameter (the same for all the items) and z denotes the latent ability.

If IRT.param = TRUE, then the parameters estimates are reported under the usual IRT parameterization, i.e.,

logit (π_i) = beta (z - beta_i^*).

The fit of the model is based on approximate marginal Maximum Likelihood, using the Gauss-Hermite quadrature rule for the approximation of the required integrals.

Value

An object of class rasch with components,

coefficients
a matrix with the parameter values at convergence. These are always the estimates of beta_i, beta parameters, even if IRT.param = TRUE.

log.Lik
the log-likelihood value at convergence.

convergence
the convergence identifier returned by optim().

hessian
the approximate Hessian matrix at convergence returned by optim().

counts
the number of function and gradient evaluations used by the quasi-Newton algorithm.

patterns
a list with two components: (i) X: a numeric matrix that contains the observed response patterns, and (ii) obs: a numeric vector that contains the observed frequencies for each observed response pattern.

GH
a list with two components used in the Gauss-Hermite rule: (i) Z: a numeric matrix that contains the abscissas, and (ii) GHw: a numeric vector that contains the corresponding weights.

max.sc
the maximum absolute value of the score vector at convergence.

constraint
the value of the constraint argument.

IRT.param
the value of the IRT.param argument.

X
a copy of the response data matrix.

control
the values used in the control argument.

na.action
the value of the na.action argument.

call
the matched call.

Warning

In case the Hessian matrix at convergence is not positive definite, try to re-fit the model using rasch(..., start.val = "random").

Note

Although the common formulation of the Rasch model assumes that the discrimination parameter is fixed to 1, rasch() estimates it. If you wish to fit the constrained version of the model, use the constraint argument accordingly. See Examples for more info.

The optimization algorithm works under the constraint that the discrimination parameter beta is always positive.

When the coefficients' estimates are reported under the usual IRT parameterization (i.e., IRT.param = TRUE), their standard errors are calculated using the Delta method.

Author(s)

Dimitris Rizopoulos d.rizopoulos@erasmusmc.nl

References

Baker, F. and Kim, S-H. (2004) Item Response Theory, 2nd ed. New York: Marcel Dekker.

Rasch, G. (1960) Probabilistic Models for Some Intelligence and Attainment Tests. Copenhagen: Paedagogiske Institute.

Rizopoulos, D. (2006) ltm: An R package for latent variable modelling and item response theory analyses. Journal of Statistical Software, 17(5), 1–25. URL http://www.jstatsoft.org/v17/i05/

See Also

coef.rasch, fitted.rasch, summary.rasch, anova.rasch, plot.rasch, vcov.rasch, GoF.rasch, item.fit, person.fit, margins, factor.scores

Examples

## The common form of the Rasch model for the 
## LSAT data, assuming that the discrimination
## parameter equals 1
rasch(LSAT, constraint = cbind(ncol(LSAT) + 1, 1))

## The Rasch model for the LSAT data under the 
## normal ogive; to do that fix the discrimination
## parameter to 1.702
rasch(LSAT, constraint = cbind(ncol(LSAT) + 1, 1.702))

## The Rasch model for the LSAT data with
## unconstraint discrimination parameter
rasch(LSAT)

## The Rasch model with (artificially created) 
## missing data
data <- LSAT
data[] <- lapply(data, function(x){
    x[sample(1:length(x), sample(15, 1))] <- NA
    x
})
rasch(data)

[Package ltm version 1.1-1 Index]