rcor.test {ltm}	R Documentation

Pairwise Associations between Items using a Correlation Coefficient

Description

Computes and tests the pairwise associations between items using a correlation coefficient

Usage

rcor.test(mat, p.adjust = FALSE, p.adjust.method = "holm", ...)

Arguments

mat
a numeric matrix or a numeric data.frame containing the manifest variables as columns.

p.adjust
logical; if TRUE the p-values are adjusted for multiple comparisons.

p.adjust.method
the method argument of p.adjust().

...
extra arguments passed to cor() and cor.test().

Value

An object of class rcor.test with components,

cor.mat
the correlation matrix.

p.values
a three column numeric matrix containing the p-values for all the combinations of items.

The print method for class rcor.test returns a square matrix in which the upper diagonal part contains the estimates of the correlation coefficients, and the lower diagonal part contains the corresponding p-values.

Note

rcor.test() is more appropriate for informal testing of association between polytomous items.

Author(s)

Dimitris Rizopoulos d.rizopoulos@erasmusmc.nl

Examples

## pairwise associations for Environment data:
rcor.test(data.matrix(Environment), method = "kendall")

## pairwise associations for independent normal random variates:
mat <- matrix(rnorm(1000), 100, 10, dimnames = list(NULL, LETTERS[1:10]))
rcor.test(mat)
rcor.test(mat, method = "kendall")
rcor.test(mat, method = "spearman")

[Package ltm version 1.1-1 Index]