summary {ltm}	R Documentation

Summary method for fitted IRT models

Description

Summarizes the fit of either grm, ltm, rasch or tpm objects.

Usage

## S3 method for class 'gpcm'
summary(object, robust.se = FALSE, ...)

## S3 method for class 'grm'
summary(object, ...)

## S3 method for class 'ltm'
summary(object, robust.se = FALSE, ...)

## S3 method for class 'rasch'
summary(object, robust.se = FALSE, ...)

## S3 method for class 'tpm'
summary(object, ...)

Arguments

object
an object inheriting from either class gpcm, either class grm, class ltm, class rasch or class tpm.

robust.se
logical; if TRUE robust estimation of standard errors is used, based on the sandwich estimator.

...
additional argument; currently none is used.

Value

An object of either class summ.gpcm, class summ.grm, class summ.ltm or class summ.rasch with components,

coefficients
the estimated coefficients' table.

Var.betas
the approximate covariance matrix for the estimated parameters; returned only in summ.ltm and summ.rasch.

logLik
the log-likelihood of object.

AIC
the AIC for object.

BIC
the BIC for object.

max.sc
the maximum absolute value of the score vector at convergence.

conv
the convergence identifier returned by optim().

counts
the counts argument returned by optim().

call
the matched call of object.

ltn.struct
a character vector describing the latent structure used in object; returned only in summ.ltm.

control
the values used in the control argument in the fit of object.

nitems
the number of items in the data set; returned only in summ.ltm and summ.rasch.

Note

For the parameters that have been constrained, the standard errors and z-values are printed as NA.

When the coefficients' estimates are reported under the usual IRT parameterization (i.e., IRT.param = TRUE in the call of either grm, ltm or rasch), their standard errors are calculated using the Delta method.

Author(s)

Dimitris Rizopoulos d.rizopoulos@erasmusmc.nl

See Also

gpcm, grm, ltm, rasch, tpm

Examples

# use Hessian = TRUE if you want standard errors
fit <- grm(Science[c(1,3,4,7)], Hessian = TRUE)
summary(fit)

## One factor model using the WIRS data;
## results are reported under the IRT
## parameterization
fit <- ltm(WIRS ~ z1)
summary(fit)

[Package ltm version 1.1-1 Index]