testEquatingData {ltm}	R Documentation

Prepares Data for Test Equating

Description

Test equating by common items.

Usage

testEquatingData(DataList, AnchoringItems = NULL)

Arguments

DataList
a list of data.frames or matrices containing common and unique items between several forms.

AnchoringItems
a data.frame or a matrix containing anchoring items for across sample equating.

Details

The purpose of this function is to combine items from different forms. Two cases are considered. Alternate Form Equating (where common and unique items are analyzed simultaneously) and Across Sample Equating (where different sets of unique items are analyzed separately based on previously calibrated anchor items).

Value

A matrix containing the common and unique items.

Author(s)

Dimitris Rizopoulos d.rizopoulos@erasmusmc.nl

References

Yu, C.-H. and Osborn Popp, S. (2005) Test equating by common items and common subjects: concepts and applications. Practical Assessment Research and Evaluation, 10(4), 1–19. URL http://pareonline.net/getvn.asp?v=10&n=4

Rizopoulos, D. (2006) ltm: An R package for latent variable modelling and item response theory analyses. Journal of Statistical Software, 17(5), 1–25. URL http://www.jstatsoft.org/v17/i05/

Examples

# Let two data-sets with common and unique items
dat1 <- as.data.frame(rmvlogis(20, cbind(c(-2, 1, 2, 1), 1)))
names(dat1) <- c("CIt2", "CIt3", "CIt4", "W")

dat2 <- as.data.frame(rmvlogis(10, cbind(c(-2, -1, 1, 2, 0.95), 1)))
names(dat2) <- c("CIt1", "CIt2", "CIt3", "CIt4", "K")

# combine in one data-set by
lisForms <- list(dat1, dat2)
testEquatingData(lisForms)

[Package ltm version 1.1-1 Index]