vcov {ltm}	R Documentation

vcov method for fitted IRT models

Description

Extracts the asymptotic variance-covariance matrix of the MLEs from either gpcm, grm, ltm, rasch or tpm objects.

Usage

## S3 method for class 'gpcm'
vcov(object, robust = FALSE, ...)

## S3 method for class 'grm'
vcov(object, ...)

## S3 method for class 'ltm'
vcov(object, robust = FALSE, ...)

## S3 method for class 'rasch'
vcov(object, robust = FALSE, ...)

## S3 method for class 'tpm'
vcov(object, ...)

Arguments

object
an object inheriting from either class gpcm, class grm, class ltm, class rasch or class tpm.

robust
logical; if TRUE the sandwich estimator is used.

...
additional arguments; currently none is used.

Value

a numeric matrix representing the estimated covariance matrix of the maximum likelihood estimates. Note that this covariance matrix is for the parameter estimates under the additive parameterization and not under the usual IRT parameterization; for more info check the Details section of grm, ltm, rasch, and tpm.

Author(s)

Dimitris Rizopoulos d.rizopoulos@erasmusmc.nl

See Also

gpcm, grm, ltm, rasch, tpm

Examples

fit <- rasch(WIRS)
vcov(fit)
sqrt(diag(vcov(fit))) # standard errors under additive parameterization

[Package ltm version 1.1-1 Index]